# Conflicts:
#	src/main/resources/static/CSS/modal.css
#	src/main/resources/templates/admin/Courses.html

+++ src/main/java/com/uet/k62/web/system/examination/restcontroller/CourseTypeController.java
@@ -5,6 +5,7 @@ import com.uet.k62.web.system.examination.model.dtos.CourseTypeRequestDTO;
 import com.uet.k62.web.system.examination.paging.PageConstant;
 import com.uet.k62.web.system.examination.service.CourseTypeService;
 import io.swagger.annotations.ApiOperation;
+import io.swagger.models.auth.In;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -35,21 +36,21 @@ public class CourseTypeController {
 
     @ApiOperation(value = "Get one course-type", response = RestBody.class)
     @GetMapping(value = "{id}")
-    public ResponseEntity getCourseType(@PathVariable BigInteger id){
+    public ResponseEntity getCourseType(@PathVariable Integer id){
         RestBody restBody = courseTypeService.getCourseType(id);
         return ResponseEntity.ok(restBody);
     }
 
     @ApiOperation(value = "Update one course-type", response = RestBody.class)
     @PutMapping(value = "{id}")
-    public ResponseEntity getCourseType(@PathVariable BigInteger id, @RequestBody CourseTypeRequestDTO dto){
+    public ResponseEntity getCourseType(@PathVariable Integer id, @RequestBody CourseTypeRequestDTO dto){
         RestBody restBody = courseTypeService.updateCourseType(dto, id);
         return ResponseEntity.ok(restBody);
     }
 
     @ApiOperation(value = "Delete a course type", response = RestBody.class)
     @DeleteMapping(value = "{id}")
-    public ResponseEntity deleteCourseType(@PathVariable BigInteger id){
+    public ResponseEntity deleteCourseType(@PathVariable Integer id){
         courseTypeService.deleteCourseType(id);
         return ResponseEntity.noContent().build();
     }
